refactor: split dialogs and loop
